+++ modules/@apostrophecms/attachment/index.js
@@ -421,7 +421,10 @@ module.exports = {
           _id: options.attachmentId ?? self.apos.util.generateId(),
           group: group.name,
           createdAt: new Date(),
-          name: self.apos.util.slugify(path.basename(fileName, path.extname(fileName))),
+          name: self.apos.util.slugify(
+            path.basename(fileName, path.extname(fileName)),
+            { stripAccents: self.apos.i18n.shouldStripAccents(req) }
+          ),
           title: self.apos.util.sortify(
             path.basename(fileName, path.extname(fileName))
           ),