+++ modules/@apostrophecms/doc/index.js
@@ -205,7 +205,10 @@ module.exports = {
       '@apostrophecms/doc-type:beforePublish': {
         testPermissions(req, info) {
           if (info.options.permissions !== false) {
-            if (!self.apos.permission.can(req, info.options.autopublishing ? 'edit' : 'publish', info.draft)) {
+            if (!self.apos.permission.can(
+              req,
+              info.options.autopublishing ? 'edit' : 'publish', info.draft
+            )) {
               throw self.apos.error('forbidden');
             }
           }
@@ -214,10 +217,19 @@ module.exports = {
       '@apostrophecms/doc-type:beforeSave': {
         ensureSlugSortifyAndUpdatedAt(req, doc, options) {
           const manager = self.getManager(doc.type);
-          manager.ensureSlug(doc);
+          manager.ensureSlug(
+            doc,
+            {
+              stripAccents: self.apos.i18n.shouldStripAccents(req)
+            }
+          );
           _.each(manager.schema, function (field) {
             if (field.sortify) {
-              doc[field.name + 'Sortified'] = self.apos.util.sortify(doc[field.name] ? doc[field.name] : '');
+              doc[field.name + 'Sortified'] = self.apos.util.sortify(
+                doc[field.name]
+                  ? doc[field.name]
+                  : ''
+              );
             }
           });
           if (options.setUpdatedAtAndBy !== false) {