+++ modules/@apostrophecms/piece-type/index.js
@@ -1146,10 +1146,10 @@ module.exports = {
       },
       // If the piece does not yet have a slug, add one based on the
       // title; throw an error if there is no title
-      ensureSlug(piece) {
+      ensureSlug(piece, { stripAccents } = {}) {
         if (!piece.slug || piece.slug === 'none') {
           if (piece.title) {
-            piece.slug = self.apos.util.slugify(piece.title);
+            piece.slug = self.apos.util.slugify(piece.title, { stripAccents });
           } else if (piece.slug !== 'none') {
             throw self.apos.error(
               'invalid',